big refactor of prediction code
seems to work, but still needs a bit more testing
